Added actions to run and publish test runs #8
